I have added a working back button and a working view asset button.
Note the page is not connected to the databse so the view asset button just links to the view asset page. i had adjusted the table size.

I just need to do the backend for the page which will require sample data and I need to make the page more responsive. I also need to add a grid view button.